Fix initial SuperUser denial
When a user starts AOKP after a full data wipe, any
requests for su will be denied. This happens because
of a missing persistent property initial setting of the
"Superuser Access" option in the SuperUser Settings.
The user then has to go to the SuperUser settings and
explicitly set the "Superuser Access" option so that
the persistent property will be set.  In CM, this property
is initialized when the user uses the "Settings->Dev Options"
menu, but as AOKP lacks this code in the global settings,
it is missing from a clean install. This update writes
the persistent property to build.prop so it is initialized
on a clean build 1st run.